assigment3/requirements.txt
psycopg==3.2.3
psycopg-binary==3.2.3
psycopg-pool==3.2.3
pydotenv==0.0.7
python-dotenv==1.0.1
typing_extensions==4.12.2
